test: make provider docs guard feature independent
